requirements-test.txt
pytest>=7.0.0  
pytest-asyncio>=0.21.0  
pytest-cov>=4.0.0  
pydantic>=1.10.0  
pytest-docker==2.0.1  
docker==6.1.3  
requests-mock==1.11.0  
behave==1.2.6  
